Switch back to configure as cmake is broken
The switch to cmake broke pkgconfig dependencies throughout,
and removed symbols that third party software is dependent on when
using loose linking.

The aforementioned issue with windres is a binutils issue, so
we must empty the RC value to ensure compilation completes
successfully.

Due to introducing harfbuzz linking in the last 32-bit build,
we'll now need to continue that and enable it correctly in this
configure-based build.

Lastly we ensure we restore freetype2-config, as it is still
required for compilation of certain software (specifically
older autotools based projects).

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>